LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/location_hash_set_empty_string-expected.txt


PASS window.location.hash is an empty string
